Rename the file to more accurately describe its
use
