+++ test/with-createreadstream-mockserver.ts
@@ -17,7 +17,7 @@ import {Datastore} from '../src';
 import * as assert from 'assert';
 
 import {startServer} from '../mock-server/datastore-server';
-import {sendNonRetryableError} from './grpc-endpoint';
+import {sendNonRetryableError} from './mock-server-tester';
 
 describe.skip('lookup', () => {
   it('should report an error to the user when it occurs', done => {
